feat(check-endpoint): add the capability to assert a status code
checkEndpointStatus represents a structure that will be common accross a lot of the assertion
mechanism. Via monadic binding, each check can be executed creating an opportunity to fail. It may
be wise to move to a Witherable instance of the current structure so errors from different
assertions can be accumulated. At this time, the checkEndpoint function performs actual useful work.
The next step is to run assertions from schemata.