chore(deps): add io-ts for schemata
io-ts is a runtime validation system. It's faster and more composable than joi, so it's really
promising for use in an actual project. The goal of this project is to check the type of some data,
initially from REST APIs, and io-ts schemata represent a very useful mechanism for both decoding and
validation of incoming data, as well as generating data for testing purposes, and finally for
producing useful typescript types.