revert(githooks): do not stash during pre-push
This doesn't work the way I dreamed it would. If there are no files to stash, you may wind up
unstashing something unexpected, or if there's nothing in the stash it'll fail entirely. There are
libraries for this, but for now it's fine. CI can be used to protect develop and master.