feat(check-endpoint): integrate schema check
Also does some type level improvements to give the returned data the real type based on the data
present. Currently, the static type derived from the schema is immediately applied and the data is
eagerly considered correct. This could represent a problem if someone made an unsafe operation
before the schema decoding. Maybe there is a way to resolve this, but it seems like somebody in the
OSS/fp-ts/io-ts community could help instead of me banging my head against a wall for a week.